refactor: 일부 필드 및 클래스 명 수정
