+++ src/main/java/com/stcom/smartmealtable/domain/social/SocialAccount.java
@@ -12,10 +12,12 @@ import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import java.time.LocalDateTime;
 import lombok.Builder;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
 @NoArgsConstructor
+@Getter
 public class SocialAccount extends BaseTimeEntity {
 
     @Id
