+++ src/main/java/com/stcom/smartmealtable/infrastructure/dto/JwtTokenResponseDto.java
@@ -1,4 +1,4 @@
-package com.stcom.smartmealtable.service.dto.token;
+package com.stcom.smartmealtable.infrastructure.dto;
 
 import lombok.Data;
 import lombok.ToString;
